+++ modules/Grades/ReportCards.php
@@ -32,10 +32,19 @@ if ( $_REQUEST['modfunc'] === 'save' )
 			PDFStop( $handle );
 		}
 		else
-			BackPrompt( _( 'No Students were found.' ) );
+		{
+			BackPrompt(
+				sprintf(
+					_( 'No %s were found.' ),
+					_( 'Grades' )
+				)
+			);
+		}
 	}
 	else
+	{
 		BackPrompt( _( 'You must choose at least one student and one marking period.' ) );
+	}
 }
 
 if ( ! $_REQUEST['modfunc'] )